requirements.txt
streamlit
qdrant-client
sentence-transformers
llama-cpp-python
python-docx
PyMuPDF